# Optimize turning a Set into a Selector