Update:

The original PR transparently optimized SelectorFromValidatedSet. However, this also changed the semantics - without a copy, the Set could mutate and impact the Selector. To avoid this, we considered just copying the Set, but it ended up being worse than the existing SelectorFromValidatedSet. Ultimately, we settled on making a new type that makes Set implement Selector, requiring users to opt-in to this potentially unsafe behavior.

New benchmarks:
```
BenchmarkSelectorFromValidatedSet
BenchmarkSelectorFromValidatedSet-48             2909588               411.9 ns/op           192 B/op          5 allocs/op
BenchmarkSetSelector
BenchmarkSetSelector-48                          9846037               121.9 ns/op             0 B/op          0 allocs/op
```

A followup PR will be made to use the new SetSelector in k/k codepaths (where applicable).

---

This PR introduces an internal change to make common hot paths in controllers optimized. It should not impact user facing behavior.

Essentially, Selectors are pretty complicated  - they are slow to evaluate and expensive (allocs) to create. A very common selector, however, is one derived from a `Set`. Constructing a set is cheap, and sometimes free (example: we use the one already allocated from Service.spec.selector), and evaluating it is also cheap.

This adds an internal specialization for Set, so that it has a custom Selector implementation.

In real world application (Istio) at scale, we found that selector computation was 5-10% of CPU in many cases. We made a similar optimization internally (by not using Selector at all) and saw this drop to 0%. This PR brings the same optimzation to all users transparently.

Fixes #112647

Benchmark:

```
$ benchstat /tmp/{old,new}
name                         old time/op    new time/op    delta
SelectorFromValidatedSet-48     397ns ± 1%     123ns ± 0%   -68.87%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
SelectorFromValidatedSet-48      192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)

name                         old allocs/op  new allocs/op  delta
SelectorFromValidatedSet-48      5.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
```


#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
NONE
```